fix fetch-prebuilt.sh

+++ Dependencies/fetch-prebuilt.sh
@@ -27,17 +27,24 @@ check_for_update() {
         if [[ "$LAST_COMMIT" != "$LATEST_COMMIT" ]]; then
             echo "Found update, downloading binaries"
             echo
-            wget -O "$1/lib$1.a" "https://github.com/SideStore/$1/releases/latest/download/lib$1-ios.a"
             wget -O "$1/lib$1-sim.a" "https://github.com/SideStore/$1/releases/latest/download/lib$1-sim.a"
-            wget -O "$1/generated.zip" "https://github.com/SideStore/$1/releases/latest/download/generated.zip"
-            echo
-            echo "Unzipping generated.zip"
-            cd "$1"
-            unzip ./generated.zip
-            mv generated/* .
-            rm generated.zip
-            rmdir generated/
-            echo "Done"
+            if [[ "$1" != "minimuxer" ]]; then
+                wget -O "$1/lib$1.a" "https://github.com/SideStore/$1/releases/latest/download/lib$1.a"
+                wget -O "$1/$1.h" "https://github.com/SideStore/$1/releases/latest/download/$1.h"
+                echo
+            else
+                wget -O "$1/lib$1-ios.a" "https://github.com/SideStore/$1/releases/latest/download/lib$1-ios.a"
+                wget -O "$1/generated.zip" "https://github.com/SideStore/$1/releases/latest/download/generated.zip"
+                echo
+                echo "Unzipping generated.zip"
+                cd "$1"
+                unzip ./generated.zip
+                mv generated/* .
+                rm generated.zip
+                rmdir generated/
+                cd ..
+                echo "Done"
+            fi
         else
             echo "Up-to-date"
         fi
